+++ examples/with-redux-observable/lib/actionTypes.js
@@ -1,0 +1,4 @@
+export const FETCH_CHARACTER_SUCCESS = 'FETCH_CHARACTER_SUCCESS'
+export const FETCH_CHARACTER_FAILURE = 'FETCH_CHARACTER_FAILURE'
+export const START_FETCHING_CHARACTERS = 'START_FETCHING_CHARACTERS'
+export const STOP_FETCHING_CHARACTERS = 'STOP_FETCHING_CHARACTERS'
