+++ examples/with-redux-observable/lib/api.js
@@ -3,15 +3,15 @@
  * They are going to be used by Epics and in getInitialProps to fetch initial data.
  */
 
-import { Observable } from "rxjs/Observable";
 import { map, catchError } from "rxjs/operators";
+import { of } from "rxjs/observable/of"
 import ajax from "universal-rx-request"; // because standard AjaxObservable only works in browser
-import { fetchCharacterSuccess, fetchCharacterFailure } from "./reducer";
+import * as actions from "./actions";
 
 export const fetchCharacter = (id, isServer) =>
   ajax({ url: `https://swapi.co/api/people/${id}` }).pipe(
-    map(response => fetchCharacterSuccess(response.body, isServer)),
+    map(response => actions.fetchCharacterSuccess(response.body, isServer)),
     catchError(error =>
-      Observable.of(fetchCharacterFailure(error.response.body, isServer))
+      of(actions.fetchCharacterFailure(error.response.body, isServer))
     )
   );
