+++ examples/with-redux-observable/lib/index.js
@@ -2,7 +2,8 @@ import { createStore, applyMiddleware } from 'redux'
 import thunkMiddleware from 'redux-thunk'
 import { createLogger } from 'redux-logger'
 import { combineEpics, createEpicMiddleware } from 'redux-observable'
-import starwarsReducer, { fetchUserEpic } from './reducer'
+import reducer from './reducer'
+import { fetchUserEpic } from './epics'
 
 const rootEpic = combineEpics(
   fetchUserEpic,
@@ -13,5 +14,5 @@ export default function initStore (initialState) {
   const logger = createLogger({ collapsed: true })  // log every action to see what's happening behind the scenes.
   const reduxMiddleware = applyMiddleware(thunkMiddleware, epicMiddleware, logger)
 
-  return createStore(starwarsReducer, initialState, reduxMiddleware)
+  return createStore(reducer, initialState, reduxMiddleware)
 };
